fixing compiler warnings with gcc 4.8: unused typedefs, return values like bool don't need be const, ...
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1204 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/spat/aabb_tree.inl
@@ -197,12 +197,12 @@ bool inline AabbTree<O, OT, SH>::intersects(const TRay& ray, TParam tMin, TParam
 	const TVector& dir = ray.direction();
 	const TVector invDir = TObjectTraits::vectorReciprocal(dir);
 #if 1
-	int stack[32];
+	size_t stack[32];
 	size_t stackSize = 0;
 	stack[stackSize++] = 0;
 	while (stackSize > 0)
 	{
-		const int index = stack[--stackSize];
+		const size_t index = stack[--stackSize];
 		LASS_SPAT_OBJECT_TREES_DIAGNOSTICS_INIT_NODE(TInfo, info);
 		LASS_ASSERT(index >= 0 && static_cast<size_t>(index) < nodes_.size());
 		const Node& node = nodes_[index];
@@ -217,7 +217,7 @@ bool inline AabbTree<O, OT, SH>::intersects(const TRay& ray, TParam tMin, TParam
 			stack[stackSize++] = index + 1;
 			continue;
 		}
-		for (int i = node.first(); i != node.last(); ++i)
+		for (size_t i = node.first(); i != node.last(); ++i)
 		{
 			LASS_SPAT_OBJECT_TREES_DIAGNOSTICS_VISIT_OBJECT;
 			if (TObjectTraits::objectIntersects(objects_[i], ray, tMin, tMax, info))
